+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - drop: specter recommendation
 - improve: module description
 - bump: version to v0.0.62
+- fork: map USB and wireless debugging switches to ADB property spoofing